Refactor: Extract BaseTrendRadarExplorer and common utilities
- Create utils.py with shared functions (parse_popularity_score,
  sanitize_prompt_input, escape_markdown_text, escape_markdown_url,
  parse_published_at, create_empty_soup)
- Create base.py with BaseTrendRadarExplorer abstract base class
- Refactor JuejinExplorer and ZhihuExplorer to inherit from base class
- Remove duplicate code from both explorer implementations
- Update content.py to use shared parse_popularity_score from utils
- Update __init__.py exports
- Fix test patches to reference new module locations
