+++ nook/services/explorers/trendradar/__init__.py
@@ -1,9 +1,17 @@
 """TrendRadar integration module."""
 
+from nook.services.explorers.trendradar.base import BaseTrendRadarExplorer
+from nook.services.explorers.trendradar.juejin_explorer import JuejinExplorer
 from nook.services.explorers.trendradar.trendradar_client import (
     TrendRadarClient,
     TrendRadarError,
 )
 from nook.services.explorers.trendradar.zhihu_explorer import ZhihuExplorer
 
-__all__ = ["TrendRadarClient", "TrendRadarError", "ZhihuExplorer"]
+__all__ = [
+    "BaseTrendRadarExplorer",
+    "JuejinExplorer",
+    "TrendRadarClient",
+    "TrendRadarError",
+    "ZhihuExplorer",
+]
