+++ nook/services/explorers/trendradar/zhihu_explorer.py
@@ -4,40 +4,12 @@
 知乎（Zhihu）のホットトピックを取得するZhihuExplorerクラスを提供します。
 """
 
-import asyncio
-import html
-import math
-import re
-import unicodedata
-from datetime import date, datetime, timezone
-from typing import Any
-
-from bs4 import BeautifulSoup
-from dateutil import parser
-
 from nook.core.config import BaseConfig
 from nook.services.base.base_feed_service import Article
-from nook.services.base.base_service import BaseService
-from nook.services.explorers.trendradar.trendradar_client import TrendRadarClient
+from nook.services.explorers.trendradar.base import BaseTrendRadarExplorer
 
 
-def _create_empty_soup() -> BeautifulSoup:
-    """空のBeautifulSoupオブジェクトを生成するファクトリ関数.
-
-    Article.soup は base_feed_service.Article の必須フィールドだが
-    ZhihuExplorerでは HTML コンテンツを使用しないため、プレースホルダーとして使用。
-    BeautifulSoupはmutableなので、毎回新しいインスタンスを返すことで
-    意図しない副作用を防止する。
-
-    Returns
-    -------
-    BeautifulSoup
-        空のBeautifulSoupインスタンス。
-    """
-    return BeautifulSoup("", "html.parser")
-
-
-class ZhihuExplorer(BaseService):
+class ZhihuExplorer(BaseTrendRadarExplorer):
     """知乎のホットトピックをTrendRadar経由で取得するExplorer.
 
     TrendRadar MCPサーバーと通信し、知乎（Zhihu）のホットトピックを
@@ -45,7 +17,7 @@ class ZhihuExplorer(BaseService):
 
     Parameters
     ----------
-    storage_dir : str, default="data"
+    storage_dir : str, default="var/data"
         データ保存ディレクトリのルートパス。
 
     Examples
@@ -54,21 +26,10 @@ class ZhihuExplorer(BaseService):
     >>> explorer.run(days=1, limit=20)
     """
 
-    # 記事ごとにGPT要約を生成するため、APIレート制限とコストを考慮して
-    # limit パラメータが指定されていない場合のデフォルト取得数 (5件)
-    # Note: limit パラメータで 1-100 の範囲で明示的に指定可能
-    TOTAL_LIMIT = 5
-
-    # GPT設定
-    GPT_TEMPERATURE = 0.3
-    GPT_MAX_TOKENS = 200
-
-    # 並列実行制限（APIレート制限対策）
-    MAX_CONCURRENT_REQUESTS = 5
-
-    # エラーメッセージ定数
-    ERROR_MSG_EMPTY_SUMMARY = "要約を生成できませんでした（空のレスポンス）"
-    ERROR_MSG_GENERATION_FAILED = "要約生成に失敗しました"
+    # プラットフォーム固有の設定
+    PLATFORM_NAME = "zhihu"
+    FEED_NAME = "zhihu"
+    MARKDOWN_HEADER = "知乎ホットトピック"
 
     def __init__(self, storage_dir: str = "var/data", config: BaseConfig | None = None):
         """ZhihuExplorerを初期化.
@@ -80,338 +41,19 @@ class ZhihuExplorer(BaseService):
         config : BaseConfig | None, default=None
             設定オブジェクト。
         """
-        # カスタムストレージディレクトリを設定に反映
-        if config is None:
-            config = BaseConfig(DATA_DIR=storage_dir)
-        else:
-            # Pydantic v2 の model_copy を使用して安全にコピー
-            config = config.model_copy(update={"DATA_DIR": storage_dir})
-
-        super().__init__("trendradar-zhihu", config=config)
-        self.client = TrendRadarClient()
-
-    async def close(self) -> None:
-        """リソースを開放."""
-        await self.client.close()
-
-    async def __aenter__(self) -> "ZhihuExplorer":
-        return self
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb) -> None:
-        await self.close()
-
-    def run(self, days: int = 1, limit: int | None = None) -> None:
-        """知乎のホットトピックを収集して保存（同期版）.
-
-        Parameters
-        ----------
-        days : int, default=1
-            何日分のデータを処理するか。
-        limit : int | None, default=None
-            取得するトピック数。Noneの場合はデフォルト値。
-        """
-        try:
-            asyncio.get_running_loop()
-        except RuntimeError:
-            asyncio.run(self._run_with_cleanup(days=days, limit=limit))
-        else:
-            raise RuntimeError(
-                "run() はイベントループ実行中には使用できません。"
-                "asyncコンテキストでは await collect(...) を使用してください。"
-            )
-
-    async def _run_with_cleanup(
-        self, days: int = 1, limit: int | None = None
-    ) -> list[tuple[str, str]]:
-        """collect実行後にクライアントをクリーンアップするラッパー.
-
-        Parameters
-        ----------
-        days : int, default=1
-            何日分のデータを処理するか。
-        limit : int | None, default=None
-            取得するトピック数。Noneの場合はデフォルト値。
-
-        Returns
-        -------
-        list[tuple[str, str]]
-            保存されたファイルパスのリスト [(json_path, md_path), ...]
-        """
-        try:
-            return await self.collect(days=days, limit=limit)
-        finally:
-            await self.close()
-
-    async def collect(
-        self,
-        days: int = 1,
-        limit: int | None = None,
-        *,
-        target_dates: list[date] | None = None,
-    ) -> list[tuple[str, str]]:
-        """知乎のホットトピックを収集して保存（非同期版）.
-
-        Parameters
-        ----------
-        days : int, default=1
-            何日分のデータを処理するか。
-        limit : int | None, default=None
-            取得するトピック数。Noneの場合はTOTAL_LIMITを使用。
-        target_dates : list[date] | None, default=None
-            対象日付のリスト（将来の拡張用）。
-
-        Returns
-        -------
-        list[tuple[str, str]]
-            保存されたファイルパスのリスト [(json_path, md_path), ...]
-        """
-        # days パラメータのバリデーション（最初にチェックして一貫した例外を返す）
-        # days != 1 は target_dates の有無に関わらずエラー（一貫性のため ValueError）
-        if days != 1:
-            raise ValueError(
-                "複数日の収集 (days > 1) はまだ実装されていません。"
-                "days=1 を指定するか、パラメータを省略してください。"
-            )
-
-        # target_dates のバリデーション
-        if target_dates is not None:
-            if len(target_dates) == 0:
-                raise ValueError(
-                    "target_dates には少なくとも1つの日付を指定してください"
-                )
-            if len(target_dates) > 1:
-                raise NotImplementedError(
-                    "複数日の収集 (len(target_dates) > 1) はまだ実装されていません。"
-                    "単一の日付を指定するか、days=1 を使用してください。"
-                )
-            # 単一日付の場合は受け入れる
-            target_date = target_dates[0]
-        else:
-            # target_dates が None の場合は今日の日付を使用
-            target_date = datetime.now(timezone.utc).date()
-
-        # limit のバリデーション
-        if limit is not None:
-            # boolはintのサブクラスなので明示的に除外
-            if (
-                not isinstance(limit, int)
-                or isinstance(limit, bool)
-                or limit < 1
-                or limit > 100
-            ):
-                raise ValueError(
-                    f"limit は 1 から 100 の整数である必要があります。指定値: {limit}"
-                )
-            effective_limit = limit
-        else:
-            effective_limit = self.TOTAL_LIMIT
-
-        # TrendRadarからデータ取得
-        news_items = await self.client.get_latest_news(
-            platform="zhihu", limit=effective_limit
+        super().__init__(
+            service_name="trendradar-zhihu",
+            storage_dir=storage_dir,
+            config=config,
         )
 
-        if not news_items:
-            self.logger.info("TrendRadarから取得したニュース項目がありません")
-            return []
-
-        # Article オブジェクトに変換
-        articles = [self._transform_to_article(item) for item in news_items]
-
-        # GPT要約を生成（並列実行でパフォーマンス向上）
-        # Semaphoreで同時実行数を制限してAPIレート制限を回避
-        sem = asyncio.Semaphore(self.MAX_CONCURRENT_REQUESTS)
-
-        async def bounded_summarize(article: Article) -> None:
-            async with sem:
-                await self._summarize_article(article)
-
-        # return_exceptions=Trueで1つの要約が失敗しても他の要約を継続
-        results = await asyncio.gather(
-            *[bounded_summarize(article) for article in articles],
-            return_exceptions=True,
-        )
-
-        # 例外チェック: _summarize_article内でキャッチしきれなかった想定外のエラーを確認
-        # Note: return_exceptions=True を使用しているため、例外はresultsに含まれる
-        # Note: CancelledError は Python 3.8+ で BaseException のサブクラスであり
-        #       Exception ではないため、ここでは捕捉されない（意図通り）
-        for i, result in enumerate(results):
-            if isinstance(result, BaseException) and not isinstance(
-                result, asyncio.CancelledError
-            ):
-                # 予期せぬエラーはログ記録のみ（再送出しない）
-                # return_exceptions=True の意図（一部の失敗を許容）と一致
-                self.logger.error(
-                    f"Unexpected error in summary generation for article '{articles[i].title}': {result}",
-                    exc_info=(type(result), result, result.__traceback__),
-                )
-                if not articles[i].summary:
-                    articles[i].summary = self.ERROR_MSG_GENERATION_FAILED
-                # Note: 処理を継続し、成功した記事は保存する
-
-        # 日付別にグループ化して保存
-        date_str = target_date.strftime("%Y-%m-%d")
-        saved_files = await self._store_articles(articles, date_str)
-
-        return saved_files
-
-    def _transform_to_article(self, item: dict[str, Any]) -> Article:
-        """TrendRadar形式のアイテムをArticleオブジェクトに変換.
-
-        Parameters
-        ----------
-        item : dict[str, Any]
-            TrendRadarから取得したニュースアイテム。
-
-        Returns
-        -------
-        Article
-            変換されたArticleオブジェクト。
-        """
-        return Article(
-            feed_name="zhihu",
-            title=str(item.get("title") or ""),
-            url=str(item.get("url") or ""),
-            text=str(item.get("desc") or ""),
-            soup=_create_empty_soup(),
-            category="hot",
-            popularity_score=self._parse_popularity_score(
-                item.get("hot") or item.get("rank", 0)
-            ),
-            published_at=self._parse_published_at(item),
-        )
-
-    def _parse_published_at(self, item: dict[str, Any]) -> datetime:
-        """記事の公開日時を解析.
-
-        Parameters
-        ----------
-        item : dict[str, Any]
-            TrendRadarアイテム。
-
-        Returns
-        -------
-        datetime
-            UTCタイムゾーン付きの解析された公開日時、または現在日時。
-        """
-        # 候補フィールド
-        candidates = ["published_at", "timestamp", "pub_date", "created_at"]
-
-        for candidate_field in candidates:
-            val = item.get(candidate_field)
-            if val is None:
-                continue
-            # Epoch timestamp handling
-            if isinstance(val, (int, float)) and not isinstance(val, bool):
-                try:
-                    return datetime.fromtimestamp(val, tz=timezone.utc)
-                except (ValueError, OverflowError, OSError):
-                    continue
-
-            # String parsing
-            try:
-                dt = parser.parse(str(val))
-                if dt.tzinfo is None:
-                    return dt.replace(tzinfo=timezone.utc)
-                return dt.astimezone(timezone.utc)
-            except (ValueError, TypeError):
-                continue
-
-        return datetime.now(timezone.utc)
-
-    def _sanitize_prompt_input(self, text: str, max_length: int = 500) -> str:
-        """プロンプト入力用のサニタイズ処理.
-
-        プロンプトインジェクション対策として、外部入力を安全に処理します。
-        - 制御文字を除去（Cc: 制御文字, Cf: フォーマット制御文字）
-        - 過度な改行を正規化
-        - 長さを制限
-
-        Parameters
-        ----------
-        text : str
-            サニタイズするテキスト。
-        max_length : int, default=500
-            最大文字数。超過分は切り捨て。
-
-        Returns
-        -------
-        str
-            サニタイズ済みテキスト。
-        """
-        if not text:
-            return ""
-        # 制御文字を除去（改行・タブは保持）
-        # Note: unicodedataベースでCc（制御文字）とCf（フォーマット制御文字）を除外
-        # JuejinExplorerと同じ実装を使用して一貫性を確保
-        sanitized = "".join(
-            char
-            for char in text
-            if char in ("\n", "\t") or unicodedata.category(char) not in ("Cc", "Cf")
-        )
-        # 連続する改行を1つに正規化
-        sanitized = re.sub(r"\n{3,}", "\n\n", sanitized)
-        # 長さ制限
-        if len(sanitized) > max_length:
-            sanitized = sanitized[:max_length] + "..."
-        return sanitized.strip()
-
-    def _parse_popularity_score(self, value: object) -> float:
-        """人気スコアを安全にパース.
-
-        Parameters
-        ----------
-        value : object
-            パースする値。
-
-        Returns
-        -------
-        float
-            パースされた人気スコア。失敗時は0.0。
-        """
-        if value is None:
-            return 0.0
-        try:
-            # 文字列の場合、カンマやプラス記号を正規化
-            if isinstance(value, str):
-                normalized = value.strip().replace(",", "")
-                if normalized.startswith("+"):
-                    normalized = normalized[1:]
-                result = float(normalized)
-            else:
-                result = float(value)
-            # NaN/Infinity は 0.0 にフォールバック
-            if not math.isfinite(result):
-                return 0.0
-            return result
-        except (ValueError, TypeError):
-            return 0.0
-
-    async def _summarize_article(self, article: Article) -> None:
-        """記事の要約を生成（破壊的変更）.
-
-        このメソッドは記事オブジェクトの `summary` フィールドを直接変更します。
-        asyncio.gather等で並列実行する場合、各Articleオブジェクトが独立している
-        ことを前提としています。同一オブジェクトを複数タスクで共有しないでください。
-
-        Parameters
-        ----------
-        article : Article
-            要約を生成する記事。
-
-        Returns
-        -------
-        None
-            戻り値なし。article.summary を直接変更します。
-        """
-        # プロンプトインジェクション対策: 外部入力をサニタイズ
+    def _get_summary_prompt(self, article: Article) -> str:
+        """GPT要約用のプロンプトを生成."""
         safe_title = self._sanitize_prompt_input(article.title, max_length=200)
         safe_url = self._sanitize_prompt_input(article.url, max_length=500)
         safe_text = self._sanitize_prompt_input(article.text or "", max_length=500)
-        # NaN/Infinity対策: 有限数でない場合は0にフォールバック
 
-        prompt = f"""以下の知乎（Zhihu）ホットトピックを日本語で詳細に要約してください。
+        return f"""以下の知乎（Zhihu）ホットトピックを日本語で詳細に要約してください。
 
 タイトル: {safe_title}
 URL: {safe_url}
@@ -430,161 +72,11 @@ URL: {safe_url}
 3. 議論の傾向 (コメントや反応から読み取れる場合)
 [議論の方向性や主要な意見について短い説明]"""
 
-        system_instruction = (
+    def _get_system_instruction(self) -> str:
+        """GPT要約用のシステム指示を取得."""
+        return (
             "あなたは中国のQ&Aプラットフォーム「知乎（Zhihu）」のトレンドを "
             "日本語で解説する専門のアシスタントです。日本のユーザーに向けて、"
             "単なる翻訳ではなく、文化的背景や議論の文脈が伝わるような"
             "具体的で情報量の多い要約を作成してください。"
         )
-
-        try:
-            # GPTClientの非同期メソッドを使用してイベントループをブロックしない
-            # Max tokensを増やして詳細な要約を許可
-            summary = await self.gpt_client.generate_async(
-                prompt=prompt,
-                system_instruction=system_instruction,
-                temperature=self.GPT_TEMPERATURE,
-                max_tokens=600,  # トークン数を増加
-            )
-            # 空の要約をフォールバック（空白のみの場合も弾く）
-            if summary and summary.strip():
-                article.summary = summary
-            else:
-                self.logger.warning(
-                    f"GPT returned empty summary for article: {article.title}"
-                )
-                article.summary = self.ERROR_MSG_EMPTY_SUMMARY
-        except asyncio.CancelledError:
-            # キャンセルは即座に再送出して呼び出し元のキャンセル処理を有効にする
-            raise
-        except Exception:
-            # 詳細はログにのみ記録（スタックトレース含む）
-            self.logger.exception(f"要約生成に失敗 (article: {article.title})")
-            # 成果物には固定メッセージのみ
-            article.summary = self.ERROR_MSG_GENERATION_FAILED
-
-    async def _store_articles(
-        self, articles: list[Article], date_str: str
-    ) -> list[tuple[str, str]]:
-        """記事をJSON/Markdown形式で保存.
-
-        Parameters
-        ----------
-        articles : list[Article]
-            保存する記事リスト。
-        date_str : str
-            日付文字列（"YYYY-MM-DD"形式）。
-
-        Returns
-        -------
-        list[tuple[str, str]]
-            保存されたファイルパス [(json_path, md_path)]
-        """
-        if not articles:
-            return []
-
-        # JSON形式にシリアライズ
-        records = []
-        for article in articles:
-            records.append(
-                {
-                    "title": article.title,
-                    "url": article.url,
-                    "feed_name": article.feed_name,
-                    "summary": article.summary,
-                    "popularity_score": article.popularity_score,
-                    "published_at": (
-                        article.published_at.isoformat()
-                        if article.published_at
-                        else None
-                    ),
-                    "category": article.category,
-                }
-            )
-
-        # JSONファイル保存
-        filename_json = f"{date_str}.json"
-        json_path = await self.save_json(records, filename_json)
-
-        # Markdown生成・保存
-        markdown = self._render_markdown(records, date_str)
-        filename_md = f"{date_str}.md"
-        md_path = await self.save_markdown(markdown, filename_md)
-
-        self.logger.info(f"{len(articles)}件の記事を保存しました: {date_str}")
-
-        return [(str(json_path), str(md_path))]
-
-    def _escape_markdown_text(self, text: str) -> str:
-        """Markdownテキスト用のエスケープ処理.
-
-        HTMLエンティティとMarkdownリンク構文を壊す文字をエスケープします。
-
-        Parameters
-        ----------
-        text : str
-            エスケープするテキスト。
-
-        Returns
-        -------
-        str
-            エスケープ済みテキスト。
-        """
-        escaped = html.escape(text)
-        return escaped.replace("[", "\\[").replace("]", "\\]")
-
-    def _escape_markdown_url(self, url: str) -> str:
-        """MarkdownリンクのURL用エスケープ処理.
-
-        Markdownリンク構文 `[text](url)` を壊す文字をエスケープします。
-
-        Parameters
-        ----------
-        url : str
-            エスケープするURL。
-
-        Returns
-        -------
-        str
-            エスケープ済みURL。
-        """
-        # Markdownリンク構文を壊す可能性のある文字をエスケープ
-        # 順序: [ ] ( ) - 括弧類をエスケープ
-        return (
-            url.replace("[", "\\[")
-            .replace("]", "\\]")
-            .replace("(", "\\(")
-            .replace(")", "\\)")
-        )
-
-    def _render_markdown(self, records: list[dict], date_str: str) -> str:
-        """記事をMarkdown形式でレンダリング.
-
-        Parameters
-        ----------
-        records : list[dict]
-            記事のリスト。
-        date_str : str
-            日付文字列。
-
-        Returns
-        -------
-        str
-            Markdownコンテンツ。
-        """
-        content = f"# 知乎ホットトピック ({date_str})\n\n"
-
-        for i, record in enumerate(records, 1):
-            title = self._escape_markdown_text(record.get("title", ""))
-            url = self._escape_markdown_url(record.get("url", ""))
-            summary = record.get("summary", "")
-            # popularity_score は型が揺れる可能性があるため、例外安全に変換
-            hot = self._parse_popularity_score(record.get("popularity_score"))
-
-            content += f"## {i}. [{title}]({url})\n\n"
-            content += f"**人気度**: {hot:,.0f}\n\n"
-            # 要約は既にMarkdown形式で構造化されているため、そのまま出力
-            content += f"**要約**:\n\n{summary}\n\n"
-            content += "---\n\n"
-
-        return content
